+++ crates/scouter_sql/src/sql/mod.rs
@@ -1,4 +1,4 @@
 pub mod postgres;
 pub mod query;
 pub mod schema;
-
+pub mod types;
