+++ crates/scouter_sql/src/sql/postgres.rs
@@ -7,6 +7,7 @@ use crate::sql::query::Queries;
 use crate::sql::schema::{
     AlertResult, FeatureResult, ObservabilityResult, QueryResult, SpcFeatureResult, TaskRequest,
 };
+use crate::sql::types::TimeInterval;
 use scouter_error::SqlError;
 use chrono::Utc;
 use cron::Schedule;
@@ -23,53 +24,6 @@ use std::result::Result::Ok;
 use std::str::FromStr;
 use tracing::{error, warn, info};
 
-
-
-pub enum TimeInterval {
-    FiveMinutes,
-    FifteenMinutes,
-    ThirtyMinutes,
-    OneHour,
-    ThreeHours,
-    SixHours,
-    TwelveHours,
-    TwentyFourHours,
-    TwoDays,
-    FiveDays,
-}
-
-impl TimeInterval {
-    pub fn to_minutes(&self) -> i32 {
-        match self {
-            TimeInterval::FiveMinutes => 5,
-            TimeInterval::FifteenMinutes => 15,
-            TimeInterval::ThirtyMinutes => 30,
-            TimeInterval::OneHour => 60,
-            TimeInterval::ThreeHours => 180,
-            TimeInterval::SixHours => 360,
-            TimeInterval::TwelveHours => 720,
-            TimeInterval::TwentyFourHours => 1440,
-            TimeInterval::TwoDays => 2880,
-            TimeInterval::FiveDays => 7200,
-        }
-    }
-
-    pub fn from_string(time_window: &str) -> TimeInterval {
-        match time_window {
-            "5minute" => TimeInterval::FiveMinutes,
-            "15minute" => TimeInterval::FifteenMinutes,
-            "30minute" => TimeInterval::ThirtyMinutes,
-            "1hour" => TimeInterval::OneHour,
-            "3hour" => TimeInterval::ThreeHours,
-            "6hour" => TimeInterval::SixHours,
-            "12hour" => TimeInterval::TwelveHours,
-            "24hour" => TimeInterval::TwentyFourHours,
-            "2day" => TimeInterval::TwoDays,
-            "5day" => TimeInterval::FiveDays,
-            _ => TimeInterval::SixHours,
-        }
-    }
-}
 
 #[derive(Debug, Clone)]
 #[allow(dead_code)]
